+++ NEWS.md
@@ -16,6 +16,10 @@
 
 * new command `bedtools`
 
+* new command `strelka`
+
+* new command `minimap2`
+
 # blit 0.2.0
 
 ## New features